Correct ops-log pagination params in API spec
The previous pagination query parameters confused pagination by id
with pagination by recorded_at. This fixes that confusion by only
paging by recorded_at. We include an anchor value for id as well
only because recorded_at values are not required to be unique, so
we can secondarily order by id after ordering by recorded_at in the
result sets. This imposes a strict, total ordering on the entire
collection which is required for pagination.