Fix handling of unauthorized requests
The ValidatingRequestHandler.prepare implementation was running even
though the AuthenticatedRequestHandler.prepare had finished the request.
This caused an OpenAPI validation failure due to the lack of security
credentials on the *unauthenticated* request.  IOW, sending a request
without a token or associated session cookie resulted in:

500 GET /projects/24/secrets (127.0.0.1): Invalid OpenAPI spec security: Security not valid for any requirement

Raising `web.Finish()` ensures that subclass prepare implementations do
not continue to process.  The result is that we send back a 401 instead
of a 500.