Add base.PaginatedRequestMixin
This class will be used to implement pagination over time-ordered
collections in the activity feed and project activity feed.  I am adding
it as a separate commit so that you can look at it in isolation before
seeing it in use.  Some of the implementation might not make sense until
you look at the following commits.

The TL;DR is that it is a class that calls a "query function" to
retrieve rows by date range and collects them into an outgoing sequence
sorted by the date.  It also is responsible for building the "token" in
the "next page" link and interpreting the query argument from the
request if present.  Part of the strangeness of the API is what makes it
more usable in the concrete classes.