Update audit columns in automations PATCH
I replaced `postgres.generate_update` with `postgres.update_entity`.
The only reason that I separated the functionality was to use the
existence of an update statement as the flag that an update was
required. Adding the last-modified columns in broke this logic. TBH, I
should have learned that using existence of something as a logic flag is
a little busted.