Fix array_agg() usage in automations endpoints
I still don't completely understand why but DISTINCT is required inside
of the two `array_agg()`s in the SELECT. If they are omitted then each
of the result arrays is forced to have the same number of items. For
example, here is the result of having an automation that is applicable
to four project types and has a single dependency.

    "applies_to": [
      "api",
      "consumers",
      "library",
      "web-app"
    ]
    "depends_on": [
      "gitlab-create-gitlab-project",
      "gitlab-create-gitlab-project",
      "gitlab-create-gitlab-project",
      "gitlab-create-gitlab-project"
    ]