Invoke project creation automations
This commit will be enhanced to implement more comprehensive rollback
behavior in future commits. Right now it simply invokes one or more
functions and deletes the project if there is a failure. The lists of
dependency and project type IDs make it a little easier to verify the
request since I do not need additional queries for translation.