Rework error handing in GitLabClient
We should pass along the status code from GitLab in the vast majority of
cases. For example, anything that makes Imbi produce a "Bad Request" on
the GitLab API shouldn't be retried... it is usually caused by incorrect
configuration somewhere in the Imbi data but that is nothing that an
automatic retry is going to fix. The only exception is that we should
never reflect a Not Authorized response from GitLab since it would be
interpreted as the Imbi user not being authorized against Imbi.