Rename PaginatedRequest classes
They are very purpose-written with iterating over database rows based on
time. The new component APIs are paginated based on name which is much
easier and, unfortunately, hard to abstract to a common base. The next
commits will introduce another set of pagination helpers that are more
flexible but not quite flexible enough for the needs of the existing
time-based ones.