Refactor SSM POST endpoint to use environment dict
This simplifies the POST endpoint to create new SSM parameters and
keeps it project-oriented instead of generic. Instead of a list of
arbitrary params to add, it now sends a single name suffix and an
object of environment -> value kv pairs. The namespace is unnecessary
because the URL path includes the project ID, which determines
the namespace. This also prevents doing any extra adds in the
same environment. It also prevents any invalid params for the
project, since now the endpoint itself appends the correct prefix
to the front of the paths.