Add component_versions to the projects index
This is a work-around for a deficiency in our FE query parsing
component. It doesn't support nested queries correctly. So... the new
`component_versions` field is a flattened version that is searchable.
This doesn't give us wildcard support (for some reason) but it does make
it possible to find projects that use a specific component as well as
projects that use a specific version of a specific component. Of course
this does induce a bit of index bloat.