Rework /projects/{id}/components response
I decided to pivot from returning numerical scores for each project
component to returning a status string instead. This commit adds the
_postprocess_item hook to the PaginatedCollectionHandler which I can use
to massage the outgoing items in the collection.

NOTE -- this commit DOES NOT include the correct logic for
distinguishing between up-to-date and out-of-date components. That will
be coming with the commit that does the score calculation.