Add operations log occurred at property
This is the user supplied value for when the event occurred at as
opposed to when the event was created at.

I also removed the ability to set or change the "recorded_by" property
since that should always come from the authorization token. The same for
the "recorded_at" property.